+++ archetypes/client/app/js/web/main.jsx
@@ -21,8 +21,9 @@ function ifAdmin(fn, ...params) {
 plugins.register()
 
 /* Admin routes */
-ui.addRoute("/admin/entities/:entity/", params => ifAdmin(ui.changeScreen, <EntitiesGrid entity={params.entity} />))
-ui.addRoute("/admin/entities/:entity/edit", params => ifAdmin(ui.changeScreen, <EntityForm entity={params.entity} />))
+ui.addRoute("/admin/entities/:entity", params => ifAdmin(ui.changeScreen, <EntitiesGrid key={params.entity} entity={params.entity} />))
+ui.addRoute("/admin/entities/:entity/:entityId", params => ifAdmin(ui.changeScreen, <EntityForm key={params.entity} entity={params.entity} entityId={params.entityId} />))
+ui.addRoute("/admin/entities/:entity/new", params => ifAdmin(ui.changeScreen, <EntityForm key={params.entity} entity={params.entity} />))
 
 
 /* Account routes */