+++ framework-widgets/src/main/java/applica/framework/widgets/forms/processors/DefaultFormProcessor.java
@@ -1,5 +1,0 @@
-package applica.framework.widgets.forms.processors;
-
-public class DefaultFormProcessor extends BaseFormProcessor {
-
-}