+++ framework-widgets/src/main/java/applica/framework/widgets/serialization/SerializationException.java
@@ -1,0 +1,25 @@
+package applica.framework.widgets.serialization;
+
+/**
+ * Created by bimbobruno on 09/01/2017.
+ */
+public class SerializationException extends Exception {
+    public SerializationException() {
+    }
+
+    public SerializationException(String message) {
+        super(message);
+    }
+
+    public SerializationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public SerializationException(Throwable cause) {
+        super(cause);
+    }
+
+    public SerializationException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}