+++ CHANGELOG.md
@@ -23,3 +23,7 @@
 - **MongoEmbedded server added (testing purposes?)**
 
 - **Mongodb Query renamed to MongoQuery to avoid confusion with framework query**
+
+- **Form processors** 
+    Form Processors can be mapped to entities using getEntityType property, otherwhise is DefaultFormProcessor.
+    Map request values changed to ObjectNode (json)